fix: made the component reusable
